requirements.txt
gunicorn>=20.1.0
numpy
tensorflow
timm
torch
torchvision
djangorestframework
django